requirements.txt
# Core dependencies
dnspython>=2.4.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
coverage>=7.3.0

# Web framework for API
flask>=3.0.0
flask-cors>=4.0.0

# Load testing
locust>=2.20.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
